feat(Typings: Support Modular Imports):
