+++ src/addons/Radio/index.d.ts
@@ -1,0 +1,4 @@
+import { CheckboxProps } from '../../modules/Checkbox/index';
+import * as React from 'react';
+
+export const Radio: React.ComponentClass<CheckboxProps>;
